+++ pages/dashboard/[id].tsx
@@ -3,6 +3,7 @@ import { useDashboardList } from '@/store/memos/useDashboardList';
 import { Dashboards } from '@/types/dashboards';
 import DashboardLayout from '@/page-layout/DashboardLayout';
 import { DashboardHeader, DashboardSidebar } from '@/components';
+import Column from '@/components/dashboard/column';
 
 export default function DashboardPage() {
   const { dashboardList } = useDashboardList();
@@ -16,13 +17,14 @@ export default function DashboardPage() {
       return String(dashboard?.id) === currentId;
     });
   }
+  console.log(dashboard);
 
   if (!dashboard) return;
   return (
     <>
       <DashboardLayout
         header={<DashboardHeader dashboard={dashboard} />}
-        main={<div>main</div>}
+        main={<Column />}
         sidebar={
           <DashboardSidebar
             dashboardList={dashboardList}
